+++ frontend/src/enums.py
